// src/main/typescript/autobahn_client/proto/message.ts
export enum MessageType {
  SERVER_STATE = 0,
  SERVER_FORWARD = 1,
  PUBLISH = 2,
  SUBSCRIBE = 3,
  UNSUBSCRIBE = 4,
  HEARTBEAT = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface AbstractMessage {
  messageType: MessageType;
}
// ...
export interface HeartbeatMessage {
  messageType: MessageType;
  uuid: string;
  topics: string[];
}

function createBaseAbstractMessage(): AbstractMessage {
  return { messageType: 0 };
}

export const AbstractMessage = {
  encode(message: AbstractMessage, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.messageType !== 0) {
      writer.uint32(8).int32(message.messageType);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): AbstractMessage {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAbstractMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.messageType = reader.int32() as any;
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AbstractMessage {
    return { messageType: isSet(object.messageType) ? messageTypeFromJSON(object.messageType) : 0 };
  },

  toJSON(message: AbstractMessage): unknown {
    const obj: any = {};
    if (message.messageType !== 0) {
      obj.messageType = messageTypeToJSON(message.messageType);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AbstractMessage>, I>>(base?: I): AbstractMessage {
    return AbstractMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AbstractMessage>, I>>(object: I): AbstractMessage {
    const message = createBaseAbstractMessage();
    message.messageType = object.messageType ?? 0;
    return message;
  },
};
// ...
export const HeartbeatMessage = {
  encode(message: HeartbeatMessage, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.messageType !== 0) {
      writer.uint32(8).int32(message.messageType);
    }
    if (message.uuid !== "") {
      writer.uint32(18).string(message.uuid);
    }
    for (const v of message.topics) {
      writer.uint32(26).string(v!);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): HeartbeatMessage {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseHeartbeatMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.messageType = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.uuid = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.topics.push(reader.string());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): HeartbeatMessage {
    return {
      messageType: isSet(object.messageType) ? messageTypeFromJSON(object.messageType) : 0,
      uuid: isSet(object.uuid) ? globalThis.String(object.uuid) : "",
      topics: globalThis.Array.isArray(object?.topics) ? object.topics.map((e: any) => globalThis.String(e)) : [],
    };
  },

  toJSON(message: HeartbeatMessage): unknown {
    const obj: any = {};
    if (message.messageType !== 0) {
      obj.messageType = messageTypeToJSON(message.messageType);
    }
    if (message.uuid !== "") {
      obj.uuid = message.uuid;
    }
    if (message.topics?.length) {
      obj.topics = message.topics;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<HeartbeatMessage>, I>>(base?: I): HeartbeatMessage {
    return HeartbeatMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<HeartbeatMessage>, I>>(object: I): HeartbeatMessage {
    const message = createBaseHeartbeatMessage();
    message.messageType = object.messageType ?? 0;
    message.uuid = object.uuid ?? "";
    message.topics = object.topics?.map((e) => e) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };